requirements.txt
fastapi
uvicorn[standard]
pydantic
openai
faiss-cpu
numpy
python-dotenv
